+++ examples/basic/src/todo-item/dto/todo-item.dto.ts
@@ -1,5 +1,5 @@
 import { GraphQLISODateTime, ID, ObjectType } from '@nestjs/graphql'
-import { CursorConnection, FilterableField } from '@ptc-org/nestjs-query-graphql'
+import { CursorConnection, FilterableField, UnPagedRelation } from '@ptc-org/nestjs-query-graphql'
 
 import { SubTaskDTO } from '../../sub-task/dto/sub-task.dto'
 import { TagDTO } from '../../tag/dto/tag.dto'
@@ -9,6 +9,11 @@ import { TagDTO } from '../../tag/dto/tag.dto'
   update: { enabled: true }
 })
 @CursorConnection('tags', () => TagDTO, {
+  update: { enabled: true },
+  remove: { enabled: true }
+})
+@UnPagedRelation('allSubTasks', () => SubTaskDTO, {
+  relationName: 'subTasks',
   update: { enabled: true },
   remove: { enabled: true }
 })