+++ examples/init-scripts/mysql/init-fetch-all-with-negative.sql
@@ -1,0 +1,5 @@
+CREATE
+USER fetch_all_with_negative;
+CREATE
+DATABASE fetch_all_with_negative;
+GRANT ALL PRIVILEGES ON fetch_all_with_negative.* TO fetch_all_with_negative;