+++ packages/query-graphql/src/loader/aggregate-relations.loader.ts
@@ -54,7 +54,7 @@ export class AggregateRelationsLoader<DTO, Relation>
         map.set(queryJson, [])
       }
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      map.get(queryJson)!.push({ ...args, index })
+      map.get(queryJson).push({ ...args, index })
       return map
     }, new Map<string, (AggregateRelationsArgs<DTO, Relation> & { index: number })[]>())
   }