+++ packages/query-graphql/src/types/connection/offset/pager/index.ts
@@ -5,4 +5,5 @@ import { OffsetPager } from './pager'
 export { OffsetPagerResult } from './interfaces'
 
 // default pager factory to plug in addition paging strategies later on.
-export const createPager = <DTO>(): Pager<DTO, OffsetPagerResult<DTO>> => new OffsetPager<DTO>()
+export const createPager = <DTO>(enableFetchAllWithNegative?: boolean): Pager<DTO, OffsetPagerResult<DTO>> =>
+  new OffsetPager<DTO>(enableFetchAllWithNegative)