+++ packages/query-typegoose/__tests__/__fixtures__/seeds.ts
@@ -88,7 +88,7 @@ export const seed = async (connection: mongoose.Connection): Promise<void> => {
       await Promise.all(
         references.map((r) => {
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          TEST_REFERENCES.find((tr) => tr._id.toString() === r._id.toString())!.testEntity = te._id
+          TEST_REFERENCES.find((tr) => tr._id.toString() === r._id.toString()).testEntity = te._id
           return r.updateOne({ $set: { testEntity: te._id } })
         })
       )
@@ -106,7 +106,7 @@ export const seed = async (connection: mongoose.Connection): Promise<void> => {
       await Promise.all(
         references.map((r) => {
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          TEST_REFERENCES_FOR_DISCRIMINATES.find((tr) => tr._id.toString() === r._id.toString())!.testEntity = tde._id
+          TEST_REFERENCES_FOR_DISCRIMINATES.find((tr) => tr._id.toString() === r._id.toString()).testEntity = tde._id
           return r.updateOne({ $set: { testEntity: tde._id } })
         })
       )